+++ src/pages/Certification/OthersAllCurrentCertification/OthersAllCurrentCertification.tsx
@@ -1,5 +1,4 @@
 import getTotalCertification from "@/apis/getTotalCertification";
-import postUserProfile from "@/apis/postUserProfile";
 import OthersProfile from "@/components/Certification/OthersAllCurrentCertification/OthersProfile/OthersProfile";
 import TotalCertification from "@/components/Certification/TotalCertification/TotalCertification";
 import Header from "@/components/Common/Header/Header";
@@ -10,6 +9,7 @@ import { useQuery } from "react-query";
 import { useParams } from "react-router-dom";
 import basicProfileImage from "@/assets/image/basic-profile-image-gray.png";
 import { QUERY_KEY } from "@/constants/queryKey";
+import { useGetUserProfile } from "@/hooks/queries/useProfileQuery";
 
 interface Data {
   totalAttempts: number;
@@ -35,15 +35,6 @@ type DayOfWeek =
   | "SATURDAY"
   | "SUNDAY";
 
-interface UserData {
-  identifier: string;
-  nickname: string;
-  fileResponse: {
-    encodedFile: "none" | string;
-  };
-  frameId: number;
-}
-
 function OthersAllCurrentCertification() {
   const { id } = useParams();
   const { userId } = useParams();
@@ -65,13 +56,7 @@ function OthersAllCurrentCertification() {
         : Promise.resolve(null),
   });
 
-  const { data: userProfile } = useQuery<UserData>({
-    queryKey: [QUERY_KEY.CERTIFICATION_USER_PROFILE, { decryptedUserId }],
-    queryFn: () =>
-      decryptedUserId
-        ? postUserProfile({ userId: parseInt(decryptedUserId) })
-        : Promise.resolve(null),
-  });
+  const { data: userProfile } = useGetUserProfile(decryptedUserId);
 
   if (!certifications) return null;
   if (!userProfile) return null;
